libstagefright: Fix accuracy of average FPS in video playback
The average FPS calculated for video playback loses
accuracy due to the truncation of the fractional parts
of each FPS value calculated during playback.

A floating type variable is now used to avoid the loss
of the fractional part from each FPS value.

(cherry picked from commit 462d97c2ca534ad44071d09c1945234b3e2dbcce)

CRs-Fixed: 331938

+++ media/libstagefright/include/AwesomePlayer.h
@@ -340,7 +340,7 @@ private:
         int64_t mFirstFrameLatencyStartUs; //first frame latency start
         int64_t mLastFrame;
         int64_t mLastFrameUs;
-        int64_t mFPSSumUs;
+        float mFPSSumUs;
         int64_t mStatisticsFrames;
         bool mVeryFirstFrame;
         int64_t mTotalTime;
